+++ src/Api/Endpoints/Zaken/Zaken.php
@@ -12,6 +12,7 @@ use Woweb\Zgw\Api\Actions\Put;
 use Woweb\Zgw\Api\Actions\Search;
 use Woweb\Zgw\Api\Actions\Show;
 use Woweb\Zgw\Api\Actions\Store;
+use Woweb\Zgw\Api\Attributes\ZgwResource;
 use Woweb\Zgw\Api\Endpoints\AbstractEndpoint;
 use Woweb\Zgw\Api\Endpoints\Zaken\Nested\ZaakBesluiten;
 use Woweb\Zgw\Api\Endpoints\Zaken\Nested\Zaakeigenschappen;
@@ -23,7 +24,6 @@ use Woweb\Zgw\Contracts\ProvidesAuditTrail;
 use Woweb\Zgw\Contracts\ReplacesResource;
 use Woweb\Zgw\Contracts\SearchesResources;
 use Woweb\Zgw\Contracts\ShowsResource;
-use Woweb\Zgw\Data\Attributes\ZgwResource;
 
 #[ZgwResource(schema: 'Zaak', component: 'zaken')]
 class Zaken extends AbstractEndpoint implements CreatesResource, DeletesResource, ListsResources, PatchesResource, ProvidesAuditTrail, ReplacesResource, SearchesResources, ShowsResource